Rework logic
